Update MyQueue.java
There was a bug in the ensureCapacity() func. the func. would copy the array as is, without considering where (on which index)  the "head" of the queue is located.
Let me demonstrate the problem using a simple quite general example:
Say the user initializes a queue of initial size 10. the user then adds 10 elements to the queue, then removes 3 elements, and finally adds 3 elements.
at this point, the array is at full capacity of 10 elements, and the head of the queue is at index 3, which is stored in the "head" variable.
So far so good. No problems up to this point. If the user would choose for example at this point, to remove 8 elements in a row, then the 7th element to be dequeued would be arr[9] and the 8th element to be dequeued would arr[0]. Just as expected, while fully preserving the FIFO principle.
But...
If instead, the user at this point, would choose to add an element, then the ensureCapacity() func. would be called, the old array would be copied as is to the new bigger array, without considering the location of the head of the queue.
The head is still "pointing" at index 3 (of the new array) but a new element has been added at index 10.
this means, that if the user chooses to remove 8 elements in a row as before, then the 7th element to be dequeued would be arr[9] just as before, but the 8th element to be dequeued would be arr[10], thus ruing the original order.
what I fixed was simple. i offset the old array index by the value of head and took its modulo arr.length. This ensures that the elements in the new array are now ordered correctly in the new array, but with the head of the queue at arr[0]. after the for loop, I then reset the value of the "head" variable to 0.